refactor: moved mobile react native to mobile/
