+++ frontend/mobile/app/(tabs)/index.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Chat from '@/components/ChatView';
+import Chat from '@/mobile/components/ChatView';
 
 export default function IndexScreen() {
   return <Chat />;
