+++ frontend/mobile/app/+not-found.tsx
@@ -1,8 +1,8 @@
 import { Link, Stack } from 'expo-router';
 import { StyleSheet } from 'react-native';
 
-import { ThemedText } from '@/components/ThemedText';
-import { ThemedView } from '@/components/ThemedView';
+import { ThemedText } from '@/mobile/components/ThemedText';
+import { ThemedView } from '@/mobile/components/ThemedView';
 
 export default function NotFoundScreen() {
   return (
