+++ frontend/mobile/assets/images/adaptive-icon.png
