+++ frontend/mobile/assets/images/partial-react-logo.png
