+++ frontend/mobile/assets/images/react-logo.png
