+++ frontend/mobile/assets/images/react-logo@3x.png
