+++ frontend/mobile/assets/images/splash-icon.png
