+++ frontend/mobile/components/Collapsible.tsx
@@ -1,10 +1,10 @@
 import { PropsWithChildren, useState } from 'react';
 import { StyleSheet, TouchableOpacity } from 'react-native';
 
-import { ThemedText } from '@/components/ThemedText';
-import { ThemedView } from '@/components/ThemedView';
-import { IconSymbol } from '@/components/ui/IconSymbol';
-import { Colors } from '@/constants/Colors';
+import { ThemedText } from '@/mobile/components/ThemedText';
+import { ThemedView } from '@/mobile/components/ThemedView';
+import { IconSymbol } from '@/mobile/components/ui/IconSymbol';
+import { Colors } from '@/mobile/constants/Colors';
 import { useColorScheme } from '@/hooks/useColorScheme';
 import React from 'react';
