+++ frontend/mobile/components/HapticTab.tsx
