+++ frontend/mobile/components/ui/IconSymbol.tsx
