+++ frontend/mobile/components/ui/TabBarBackground.ios.tsx
