+++ frontend/mobile/components/ui/TabBarBackground.tsx
