+++ frontend/mobile/scripts/reset-project.js
